Added code sufficient to pass failing end to end test which required token of job processing. RockJob now available externally for evaluation. Update to dev notes - changed todo's to E2E test, added new Todo section

+++ RockProcessing/RockProcessing.Model/RockJob.cs
@@ -13,7 +13,7 @@ namespace RockProcessing.Model {
 		private bool _complete;
 		private double _currentWeight;
 
-		public RockJob(RockType rockType, int weight, IRockJobMonitor monitor)
+		public RockJob(RockType rockType, double weight, IRockJobMonitor monitor)
 		{
 			_rockType = rockType;
 			_originWeight = weight;
